requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
email-validator==2.1.0.post1
PyMySQL==1.1.0
cryptography
python-dotenv
openpyxl
resend
psycopg2-binary
supabase
python-dotenv
xhtml2pdf
